End Admin App

+++ projects/admin/src/app/dashboard/manage-users/components/users/users.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { SpinnerService } from 'projects/shared/spinner.service';
-import { UsersService } from '../../services/users.service';
+import { changeStatus, UsersService } from '../../services/users.service';
 import { ToastrService } from 'ngx-toastr';
 
 @Component({
@@ -16,50 +16,79 @@ export class UsersComponent implements OnInit {
   itemsPerPageOptions: number[] = [1, 2, 5, 10];
   totalItems: any;
   timeOutId: any
-  constructor(public spinnerService: SpinnerService, private userService: UsersService, private toastr: ToastrService) { }
+  constructor(public spinnerService: SpinnerService, private userService: UsersService, private toastr: ToastrService) {
+    this.getUserFromSubject();
+  }
   ngOnInit(): void {
-    this.getUserData();
+    this.getUser();
   }
 
 
-
-  getUserData() {
+  getUser(name: string = '') {
     const model = {
       page: this.page,
       limit: this.itemsPerPage,
-      name: '',
+      name: name,
     }
-    this.userService.getAllUser(model).subscribe((res: any) => {
-      this.totalItems = res.totalItems;
-      this.dataSource = res.users
+    this.userService.getUserData(model);
+  }
+
+  getUserFromSubject() {
+    this.userService.userData.subscribe((res: any) => {
+      this.dataSource = res.data;
+      this.totalItems = res.total;
     })
   }
 
-  deleteUser(id: any) {
-    this.userService.deleteUser(id).subscribe((res) => {
-      this.toastr.success("User Deleted Successfully");
-      this.getUserData();
-    })
+
+  deleteUser(id: any, index: number) {
+    if (this.dataSource[index].assignedTasks > 0) {
+      this.toastr.error("You Can't Delete This User Until Finish Tasks")
+    }
+    else {
+      this.userService.deleteUser(id).subscribe((res) => {
+        this.toastr.success("User Deleted Successfully");
+        this.page = 1;
+        this.getUser();
+      })
+    }
   }
 
   changePage(event: any) {
     console.log(event);
     this.page = event;
-    this.getUserData();
+    this.getUser();
   }
 
   onItemsPerPageChange() {
     this.page = 1;
-    this.getUserData();
+    this.getUser();
   }
 
   search(event: any) {
-    // this.model['name'] = event.value;
+
     this.page = 1;
     clearTimeout(this.timeOutId)
     this.timeOutId = setTimeout(() => {
-      this.getUserData();
+      this.getUser(event.target.value);
     }, 1000);
+  }
+
+  changeUserStatus(status: string, id: string, index: number) {
+    const Model: changeStatus = {
+      id: id,
+      status: status
+    }
+    if (this.dataSource[index].assignedTasks > 0) {
+      this.toastr.error("You Can't Update User Status Until Finish Tasks")
+    }
+    else {
+      this.userService.changeStatus(Model).subscribe((res) => {
+        this.toastr.success("User Status Updated Successfully");
+        this.page = 1;
+        this.getUser();
+      })
+    }
   }
 
 }
